+++ src/proxy-storage.js
@@ -124,15 +124,17 @@ class WebStorage {
    *
    * @param  {string} key: keyname of the storage
    * @param  {any} value: data to save in the storage
+   * @param  {object} options: additional options for cookieStorage
    * @return {void}
    *
    * @memberOf WebStorage
    */
-  setItem(key, value) {
+  setItem(key, value, options) {
     checkEmpty(key);
-    executeInterceptors('setItem', key, value);
+    this[key] = value;
+    executeInterceptors('setItem', key, value, options);
     value = JSON.stringify(value);
-    this.__storage.setItem(key, value);
+    this.__storage.setItem(key, value, options);
   }
   /**
    * Retrieves a value by its key name.
@@ -158,6 +160,7 @@ class WebStorage {
    */
   removeItem(key) {
     checkEmpty(key);
+    delete this[key];
     executeInterceptors('removeItem', key);
     this.__storage.removeItem(key);
   }
@@ -170,7 +173,18 @@ class WebStorage {
    */
   clear() {
     executeInterceptors('clear');
+    Object.keys(this).forEach((key) => delete this[key]);
     this.__storage.clear();
+  }
+  /**
+   * Gets the number of data items stored in the Storage object.
+   *
+   * @readonly
+   *
+   * @memberOf WebStorage
+   */
+  get length() {
+    return Object.keys(this).length;
   }
   /**
    * Adds an interceptor to a WebStorage method
@@ -245,6 +259,56 @@ const configStorage = {
 /**
  * @private
  *
+ * Determines whether a value is a plain object
+ *
+ * @param  {any} value: the object to test
+ * @return {boolean}
+ */
+function isObject(value) {
+  return Object.prototype.toString.call(value) === '[object Object]';
+}
+
+/**
+ * @private
+ *
+ * Allows add or subtract timestamps to the current date or to a specific date.
+ * @param  {object} options: It contains the timestamps to add or remove to the date, and have the following properties:
+ *         - {Date} date: if provided, the timestamps will affect this date, otherwise a new current date will be used.
+ *         - {number} hours: hours to add/subtract
+ *         - {number} days: days to add/subtract
+ *         - {number} months: months to add/subtract
+ *         - {number} years: years to add/subtract
+ * @return {Date}
+ */
+function setTimestamp(options = {}) {
+  const opt = Object.assign({}, options);
+  let d = opt.date instanceof Date ? opt.date : new Date();
+  if (+opt.hours) d.setHours(d.getHours() + opt.hours);
+  if (+opt.days) d.setDate(d.getDate() + opt.days);
+  if (+opt.months) d.setMonth(d.getMonth() + opt.months);
+  if (+opt.years) d.setFullYear(d.getFullYear() + opt.years);
+  return d;
+}
+
+/**
+ * @private
+ *
+ * Builds the expiration part for the cookie
+ *
+ * @param  {Date|object} date: the expiration date. See `setTimestamp(options)`
+ * @return {string}
+ */
+function buildExpirationString(date) {
+  let expires = (date instanceof Date ?
+    setTimestamp({date}) :
+    setTimestamp(date)
+  );
+  return `; expires=${expires.toUTCString()}`;
+}
+
+/**
+ * @private
+ *
  * Alias for the default cookie storage associated with the current document.
  *
  * @Reference
@@ -269,15 +333,14 @@ const $cookie = {
  */
 function cookieStorage() {
   const api = {
-    setItem(key, value, days, path = '/') {
-      let expires = '';
-      if (days) {
-        const date = new Date();
-        days = days * 24 * 60 * 60 * 1000;
-        date.setTime(date.getTime() + days);
-        expires = `; expires=${date.toUTCString()}`;
+    setItem(key, value, options) {
+      let expires = '', cookie;
+      options = Object.assign({path: '/'}, options);
+      if (isObject(options.expires) || options.expires instanceof Date) {
+        expires = buildExpirationString(options.expires);
       }
-      $cookie.set(`${key}=${encodeURIComponent(value)}${expires}; path=${path}`);
+      cookie = `${key}=${encodeURIComponent(value)}${expires}; path=${options.path}`;
+      $cookie.set(cookie);
     },
 
     getItem(key) {
@@ -293,7 +356,7 @@ function cookieStorage() {
     },
 
     removeItem(key) {
-      api.setItem(key, '', -1);
+      api.setItem(key, '', {expires: {days: -1}});
     },
 
     clear() {